// src/events/event-types.ts
export enum EventType {
  CLIENT_CREATED = "client.created",
  CLIENT_UPDATED = "client.updated",

  PRODUCT_CREATED = "product.created",
  PRODUCT_UPDATED = "product.updated",

  SALESORDER_CREATED = "salesorder.created",
  SALESORDER_UPDATED = "salesorder.updated",

  INVOICE_CREATED = "invoice.created",

  INVENTORY_UPDATED = "inventory.updated",

  WORKORDER_CREATED = "workorder.created",
}